# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Тестовое задание на позицию Junior Backend Developer",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Access and Refresh tokens",
                "operationId": "create-tokens",
                "parameters": [
                    {
                        "description": "Json with user id",
                        "name": "UserId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Access and Refresh tokens",
                "operationId": "refresh-tokens",
                "parameters": [
                    {
                        "description": "Json with user id",
                        "name": "UserId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Request"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access token",
                        "name": "Access",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "Refresh",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "handler.Request": {
            "type": "object",
            "properties": {
                "UserId": {
                    "type": "string"
                }
            }
        }
    }
}